Spell Import: Create Saving Throw Spell Type
This fixes the import to save spells on the character sheet as Saving
Throw based rather than attack base, if they are listed as such.

If a spell requires a saving throw and an attack, it defaults to attack.
The logic in this case will still add a saving throw.

I tested this with the Ice Dagger spell.