Fix duplicate action features
When a class/feat/racial feature exists as an action, it appears twice
in the DDB input data. Filter these out of the list of passive abilities
so they only appear as an active ability action and not as a passive
ability entry. e.g. Barbarian Rage, Half-Orc Relentless Endurance.